FO-2388 hent oppfolgingsstatus fra Arena dersom bruker under oppfolging og 404 fra veilarbarena
